chore: edit type of Point(Integer->BigDecimal) to accurate caculation

+++ src/main/java/likelion/babsim/web/member/MemberResDTO.java
@@ -3,6 +3,7 @@ package likelion.babsim.web.member;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -13,10 +14,10 @@ public class MemberResDTO {
     private String email;
     private Integer job;
     private List<Long> allergies;
-    private Integer point;
+    private BigDecimal point;
 
     @Builder
-    public MemberResDTO(String id, String name, String img, String email, Integer job, List<Long> allergies, Integer point) {
+    public MemberResDTO(String id, String name, String img, String email, Integer job, List<Long> allergies, BigDecimal point) {
         this.id = id;
         this.name = name;
         this.img = img;
